Allow program to be run as a script
